The code is shortened; matching algorithm and list improved.
If the word is not found in the dataset, the program automatically cross-references it to the "matched" list of lower-cased data keys elements (numbered choices). Each choice in the list is then overridden by its original form present in the dataset (so not necessarily lower-cased). The list is presented to the user with words in their original format. Last part of code is aggregated into an if statement nested into another.